Getting much closer to 1.2 release!
git-svn-id: http://svn.cmeerw.net/openwatcom@2930 9848d260-d8c2-4c77-ac60-93a84d57cbad

+++ branch/release/bld/f77/rt/rmemmgr.c
@@ -24,24 +24,18 @@
 *
 *  ========================================================================
 *
-* Description:  WHEN YOU FIGURE OUT WHAT THIS FILE DOES, PLEASE
-*               DESCRIBE IT HERE!
+* Description:  Run time memory manager.
 *
 ****************************************************************************/
 
-
-//
-//  RMEMMGR : Run time memory manager
-//
 
 #include "ftnstd.h"
 #if _OPT_CG == _OFF
 #include "bglobal.h"
 #endif
 
-extern  void            *_SysMemAlloc(int);
+extern  void            *_SysMemAlloc(size_t);
 extern  void            _SysMemFree(void *);
-
 
 void            *RMemAlloc( int size ) {
 //======================================
